backend/pyproject.toml
[project]
name = "bendbionics"
version = "0.1.0"
description = "Soft Robot Simulation Platform - Backend API"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "click==8.3.1",
    "fastapi==0.121.3",
    "h11==0.16.0",
    "httptools==0.7.1",
    "idna==3.11",
    "pydantic==2.12.4",
    "pydantic-settings==2.12.0",
    "python-dotenv==1.2.1",
    "PyYAML==6.0.3",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.44",
    "sqlmodel==0.0.27",
    "starlette>=0.40.0,<0.51.0",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
    "httpx==0.28.1",
    "python-jose[cryptography]==3.5.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.3.0",
    "python-multipart==0.0.20",
    "alembic==1.17.2",
    "psycopg2-binary==2.9.11",
    "asyncpg==0.30.0",
    "python-decouple==3.8",
    "numpy==1.26.4",
    "gunicorn==21.2.0",
    "aiofiles==23.2.1",
    "aiohttp==3.13.2",
    "ruff==0.14.6",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "pytest-asyncio==0.26.0",
    "pytest-cov==6.3.0",
    "coverage==7.12.0",
    "black",
    "pylint",
    "mypy",
    "isort",
    "safety",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = [
    "E203",  # whitespace before ':'
    "S101",  # use of assert
    "PLR0913",  # too many arguments
    "PLR0912",  # too many branches
    "PLR0915",  # too many statements
    "C901",  # too complex
    "UP045",  # non-pep604-annotation-optional
    "UP006",  # non-pep585-annotation
    "UP035",  # deprecated-import
    "COM812",  # missing-trailing-comma
    "RUF022",  # unsorted-dunder-all
    "B905",  # zip-without-explicit-strict
    "PLR5501",  # collapsible-else-if
    "RUF010",  # explicit-f-string-type-conversion
    "B008",  # do not perform function call in argument defaults (FastAPI Depends pattern)
    "B904",  # within except clause, raise exceptions with raise ... from err
    "DTZ003",  # datetime.utcnow() used
    "N806",  # variable name should be lowercase
    "PLR2004",  # magic value used in comparison
    "PLC0415",  # import should be at top-level
    "ERA001",  # found commented-out code
    "S105",  # possible hardcoded password
    "ARG001",  # unused function argument
    "ARG002",  # unused method argument
    "TRY004",  # prefer TypeError exception for invalid type
    "TRY300",  # consider moving statement to else block
    "S110",  # try-except-pass detected
    "SIM102",  # use single if statement instead of nested if
    "SIM117",  # use single with statement with multiple contexts
    "PLR0911",  # too many return statements
    "N803",  # argument name should be lowercase
    "RUF006",  # store reference to return value of asyncio.create_task
    "INP001",  # file is part of implicit namespace package
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "PLR2004", "S106", "S108"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=40"
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
    "*/.venv/*",
    "*/.env/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@\\(abc\\\\.\\)\\?abstractmethod",
    "def main\\(\\):",
    "if TYPE_CHECKING:",
    "from __future__ import annotations"
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.3.0",
]

